Simplify basemap choices
This removes the Esri streets basemap in favor of Topo (which actually we call 'Streets'). Having many basemaps presents a design challenge: the map is a huge part of our site and visually we want colors of elements (header, footer, etc) to coordinate with the map. We will stick with only these two basemaps until a strong use case for more is found. This simplifies our UI.